+++ src/lib/FormData.php
@@ -3,15 +3,28 @@
 namespace NitroXy\PHPForms;
 
 class FormData {
-	public function __construct(array $data = null){
-		if ( $data === null ) return;
+	protected $data;
 
-		foreach ( $data as $key => $value ){
-			$this->$key = $value;
+	public function __construct($data){
+		if ( $data === null ){
+			$this->data = new \stdClass();
+			return;
+		} else if ( is_array($data) ){
+			$this->data = (object)$data;
+		} else {
+			$this->data = $data;
 		}
 	}
 
-	public function has_errors(){
+	public function getValueFor($key){
+		return isset($this->data->$key) ? $this->data->$key : false;
+	}
+
+	public function getId(){
+		return $this->getValueFor('id');
+	}
+
+	public function getErrorFor($field){
 		return false;
 	}
 };